support video play on storybook by removing origin
