+++ src/YoutubeAtom/YoutubeAtom.stories.tsx
@@ -18,17 +18,13 @@ export const DefaultStory = (): JSX.Element => {
             `}
         >
             <YoutubeAtom
-                format={{
-                    pillar: Pillar.News,
-                    design: Design.Article,
-                    display: Display.Standard,
-                }}
                 videoMeta={{
                     assetId: '-ZCvZmYlQD8',
                     mediaTitle:
                         "Rayshard Brooks says US justice system treats people 'like animals' in interview before his death",
                 }}
                 duration={252}
+                isStory={true}
             />
         </div>
     );
@@ -43,11 +39,6 @@ export const WithOverlayImage = (): JSX.Element => {
             `}
         >
             <YoutubeAtom
-                format={{
-                    pillar: Pillar.News,
-                    design: Design.Article,
-                    display: Display.Standard,
-                }}
                 videoMeta={{
                     assetId: '3jpXAMwRSu4',
                     mediaTitle:
@@ -55,6 +46,7 @@ export const WithOverlayImage = (): JSX.Element => {
                 }}
                 duration={252}
                 overlayImage="https://i.guim.co.uk/img/media/4b3808707ec341629932a9d443ff5a812cf4df14/0_309_1800_1081/master/1800.jpg?width=1200&height=630&quality=85&auto=format&fit=crop&overlay-align=bottom%2Cleft&overlay-width=100p&overlay-base64=L2ltZy9zdGF0aWMvb3ZlcmxheXMvdGctZGVmYXVsdC5wbmc&enable=upscale&s=aff4b8255693eb449f13070df88e9cac"
+                isStory={true}
             />
         </div>
     );
