+++ src/YoutubeAtom/YoutubeAtom.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import { YoutubeOverlay } from './YoutubeOverlay';
 import { MaintainAspectRatio } from './MaintainAspectRatio';
-import { YoutubeAtomType } from '../types';
 
 type EmbedConfig = {
     adsConfig: {
@@ -43,9 +42,30 @@ const constructQuery = (query: { [key: string]: any }): string =>
         })
         .join('&');
 
+type YoutubeMeta = {
+    assetId: string;
+    mediaTitle: string;
+    id?: string;
+    channelId?: string;
+    duration?: number;
+    posterSrc?: string;
+    height?: string;
+    width?: string;
+};
+
+type YoutubeAtomType = {
+    videoMeta: YoutubeMeta;
+    overlayImage?: string;
+    adTargeting?: AdTargeting;
+    height?: number;
+    width?: number;
+    title?: string;
+    duration?: number; // in seconds
+    isStory?: boolean; // workaround to remove origin in src to allow play in storybook
+};
+
 // Note, this is a subset of the CAPI MediaAtom essentially.
 export const YoutubeAtom = ({
-    format,
     videoMeta,
     overlayImage,
     adTargeting,
@@ -53,6 +73,7 @@ export const YoutubeAtom = ({
     width = 460,
     title = 'YouTube video player',
     duration,
+    isStory = false,
 }: YoutubeAtomType): JSX.Element => {
     const embedConfig =
         adTargeting && JSON.stringify(buildEmbedConfig(adTargeting));
@@ -64,7 +85,11 @@ export const YoutubeAtom = ({
                     width={width}
                     height={height}
                     id={videoMeta.assetId}
-                    src={`https://www.youtube.com/embed/${videoMeta.assetId}?embed_config=${embedConfig}&enablejsapi=1&origin=https://www.theguardian.com&widgetid=1&modestbranding=1`}
+                    src={`https://www.youtube.com/embed/${
+                        videoMeta.assetId
+                    }?embed_config=${embedConfig}&enablejsapi=1${
+                        isStory ? '' : '&origin=https://www.theguardian.com'
+                    }&widgetid=1&modestbranding=1`}
                     // needed in order to allow `player.playVideo();` to be able to run
                     // https://stackoverflow.com/a/53298579/7378674
                     allow="autoplay"
